Tests LoyaltyProgram. Fix side effects of static fields. Replace static store and id/sequnceNumber fields with StoreSource and ValueGenerators

+++ Microservices.NetCore.LoyaltyProgram.EventConsumer/Event.cs
@@ -2,7 +2,7 @@ namespace Microservices.NetCore.LoyaltyProgram.EventConsumer;
 
 public struct Event
 {
-    public long SequenceNumber { get; set; }
+    public int SequenceNumber { get; set; }
     public string Name { get; set; }
     public object Content { get; set; }
 }
